feat: move buttons customization to folder
