+++ extensions/measurement-tracking/src/getCustomizationModules.ts
@@ -1,0 +1,14 @@
+import viewportButtons from './customizations/viewportButtons';
+
+function getCustomizationModule() {
+  return [
+    {
+      name: 'default',
+      value: {
+        ...viewportButtons,
+      },
+    },
+  ];
+}
+
+export default getCustomizationModule;
